+++ packages/in_app_purchase/android/src/main/java/io/flutter/plugins/inapppurchase/MethodCallHandlerImpl.java
@@ -73,7 +73,7 @@ class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
         isReady(result);
         break;
       case InAppPurchasePlugin.MethodNames.START_CONNECTION:
-        startConnection((int) call.argument("handle"), result);
+        startConnection((int) call.argument("handle"),(boolean)call.argument("enablePendingPurchases"), result);
         break;
       case InAppPurchasePlugin.MethodNames.END_CONNECTION:
         endConnection(result);
@@ -236,9 +236,9 @@ class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
         });
   }
 
-  private void startConnection(final int handle, final MethodChannel.Result result) {
+  private void startConnection(final int handle, final boolean enablePendingPurchases, final MethodChannel.Result result) {
     if (billingClient == null) {
-      billingClient = billingClientFactory.createBillingClient(applicationContext, methodChannel);
+      billingClient = billingClientFactory.createBillingClient(applicationContext, methodChannel, enablePendingPurchases);
     }
 
     billingClient.startConnection(